# Use a bigger output buffer for writing to gzip and standard files