--- a/src/xopen/__init__.py
+++ b/src/xopen/__init__.py
@@ -35,6 +35,8 @@ from typing import Optional, Union, TextIO, AnyStr, IO, List, Set
 
 from ._version import version as __version__
 
+BUFFER_SIZE = max(io.DEFAULT_BUFFER_SIZE, 128 * 1024)
+
 
 try:
     from isal import igzip, isal_zlib  # type: ignore
@@ -975,7 +977,7 @@ def xopen(  # noqa: C901  # The function is complex, but readable.
     elif detected_format == "bz2":
         opened_file = _open_bz2(filename, mode, threads, **text_mode_kwargs)
     else:
-        opened_file = open(filename, mode, **text_mode_kwargs)
+        opened_file = open(filename, mode, **text_mode_kwargs, buffering=BUFFER_SIZE)
 
     # The "write" method for GzipFile is very costly. Lots of python calls are
     # made. To a lesser extent this is true for LzmaFile and BZ2File. By