--- a/src/xopen/__init__.py
+++ b/src/xopen/__init__.py
@@ -35,6 +35,7 @@ from typing import Optional, Union, TextIO, AnyStr, IO, List, Set
 
 from ._version import version as __version__
 
+# 128K buffer size also used by cat, pigz etc. It is faster than the 8K default.
 BUFFER_SIZE = max(io.DEFAULT_BUFFER_SIZE, 128 * 1024)
 
 
@@ -988,9 +989,7 @@ def xopen(  # noqa: C901  # The function is complex, but readable.
         isinstance(opened_file, (gzip.GzipFile, bz2.BZ2File, lzma.LZMAFile))
         and "w" in mode
     ):
-        # 128K buffer size also used by cat, pigz etc. It is faster than
-        # the 8K default.
         opened_file = io.BufferedWriter(
-            opened_file, buffer_size=128 * 1024  # type: ignore
+            opened_file, buffer_size=BUFFER_SIZE  # type: ignore
         )
     return opened_file