Before:
```
Benchmark #1: python dnaio_read_and_write.py ~/test/big2.fastq ramdisk/output.fastq.gz
  Time (mean ± σ):      7.585 s ±  0.050 s    [User: 7.306 s, System: 0.278 s]
  Range (min … max):    7.512 s …  7.678 s    10 runs
```

After
```
Benchmark #1: python dnaio_read_and_write.py ~/test/big2.fastq ramdisk/output.fastq.gz
  Time (mean ± σ):      7.270 s ±  0.064 s    [User: 6.989 s, System: 0.279 s]
  Range (min … max):    7.129 s …  7.351 s    10 runs
```
Reading and writing dnaio without gzip is about 2 seconds. So the gzip compression time goes from 5.6 to 5.3. That is not too shabby.