@rhpvorderman 2022-01-14
Also, this begs the question what default buffers python's gzip is using. Probably an 8k buffer as well.
Lately on python-isal I have updated all readers on the python-isal IgzipReader to be 128k. And finally finally, I got close to the speed of the C application igzip. Just 6% overhead. See: https://github.com/pycompression/python-isal/pull/106