@rhpvorderman 2022-01-14
And probably we should checkout suboprocess.PIPE as well. That will probably not be a 128K buffer too.