@marcelm 2022-01-14
Nice! Let me know if you think this justifies another release.