@rhpvorderman 2022-01-17
I just tested the performance of updating Popen's bufsize argument with 128K. On reading the performance was the same. On writing the performance was very slightly better. (70ms saved). But it doesn't seem worth it to mess with it. I think the buffer sizes are fine now.